pyproject.toml
[tool.poetry]
name = "csst.analyzer"
version = "4.0.0"
description = ""
authors = ["Joe Kern <jkern34@gatech.edu>", "Mona Amrihesari <mamrihesari3@gatech.edu>"]
readme = "README.md"
packages = [{include = "csst"}]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.4.4"
numpy = "^1.23.2"
seaborn = "^0.12.0"
matplotlib = "^3.5.3"
pydantic = "^1.10.2"
sqlalchemy = "^2.0.11"
psycopg = {version = "^3.1.17", extras = ["binary", "pool"]}
python-dotenv = { version = "^1.0.0", optional = false }
jupyterlab = "^4.0.9"
alembic = "^1.13.1"
scipy = "^1.12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
black = "^23.1.0"
sphinx = "^6.1.3"
sphinxcontrib-napoleon = "^0.7"
sphinx-rtd-theme = "^1.2.0"
enum-tools = "^0.9.0.post1"
sphinx-jinja2-compat = "^0.2.0"
sphinx-toolbox = "^3.4.0"
ruff = "^0.1.2"
sqlalchemy-utils = "^0.41.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
